inherent db table column from parent class

+++ Spring-10-ORM-Table-Creation/src/main/java/com/cydeo/entity/Employee.java
@@ -1,13 +1,14 @@
 package com.cydeo.entity;
 
 import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 
 @Entity
 public class Employee {
 
     @Id
-    private int id;
+    private Long id;
     private String name;
